HODS-401 | Removed redundant ScopesConfig
